requirements.txt
django>=6.0
django-filter>=24.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-health-check>=4.1.0
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
pandas>=2.2
